Fixed skipping cloud task manager unit tests when not on Google Cloud.
