+++ loom/common/helper.py
@@ -55,17 +55,3 @@ def get_null_logger(level=None, name=None):
 
     return logger
 
-def on_gcloud_vm():
-    """ Determines if we're running on a GCE instance."""
-    import requests
-    r = None
-    try:
-        r = requests.get('http://metadata.google.internal')
-    except ConnectionError:
-        return False
-
-    try:
-        if r.headers['Metadata-Flavor'] == 'Google' and r.headers['Server'] == 'Metadata Server for VM':
-            return True
-    except KeyError:
-        return False
